Fix test CSR versions for OpenSSL 3.4 change
The only valid CSR version is version 1, which is encoded
with "csr.version = 0". Versions of OpenSSL prior to 3.4
would allow other versions to be set / versions to be unset
but OpenSSL 3.4 is strict about only allowing version 1.

This change updates the unit tests to use version 1 in
all CSRs generated as part of testing.  It also adds
the start of a script to help regenerate the
spec/fixtures/ssl directory contents.