(maint) Add a --trace to the gha checks rake calls
poking Rake to try and figure out why the require 'packaging' is urping
in the gha run. Turns out that was due to a lack of tags in my fork and
packaging apparently performing some git describes on require.

Running rake with --trace is probably generally helpful, particularly if
the packaging gem is going to be fiddly, so I'll leave it there.