Add MacOS signing
This utilizes changes in Vanagon to enable signing locally, instead of shipping files to a signing server. Details for how to set this up given the appropriate certs/keys can be found on the wiki.

With the changes from https://github.com/OpenVoxProject/vanagon/pull/13, when the VANAGON_FORCE_SIGNING env var is set, vanagon will attempt the signing flow as noted in that PR. Otherwise, it will skip signing entirely. In order for signing to succeed, you must run the build on a VM that contains:
- A keychain created by root that contains the Developer Application and Installer identities (cert + key), and a Notarization profile.
- SIGNING_KEYCHAIN set to the name/location of the keychain
- SIGNING_KEYCHAIN_PW set to the password to unlock the keychain
- APPLICATION_SIGNING_CERT set to the description of the application signing identity
- INSTALLER_SIGNING_CERT set to the description of the installer signing identity
- NOTARY_PROFILE set to the name of the profile set up with a user account tied to the above signing identities

At some point, we may move all this out of vanagon and put it back in here, but since this should never really change and the only thing we're signing right now is the agent for MacOS, this works well enough.