Revert "Deduplicate where we get the version from"
This reverts commit bf87d142f2643e05c4c5564ad6eb732e8331cef7.

When building a gem, this file is processed by the gem(1) command and
the installed file is the result of a bunch of substitutions.

However, when building AIO packages, the file is copied as is, and
a relative path will not work properly.  This result in a flood of
warnings about invalid gemspec:

```
Invalid gemspec in [/opt/puppetlabs/puppet/lib/ruby/gems/3.2.0/specifications/puppet-8.19.0.gemspec]: cannot load such file -- /opt/puppetlabs/puppet/lib/ruby/gems/3.2.0/specifications/lib/puppet/version
```