(gh-101) Configure agent server to master fqdn instead of beaker.hostname
When running the acceptance suite, each agent node was getting its
puppet.conf server set to the Beaker master node's hostname value as set
in the hosts.yaml. The values in hosts.yaml are somewhat arbitrary, and
the puppetserver cert is being configured with dns_alt_names matching
fqdn. This patch changes the agent server setting to also be fqdn so
they match.
(determined in this test by `facter fqdn`)