(gh-101) Get i18n tests passing be ensuring locales present
The i18n tests were failing on almalinux and debian because of locale
pre-requisites.

Almalinux 8/9 does not have the glibc-all-langpacks installed, unlike
rocky, and only has the en_* locales by default. Debian has only the
base C.UTF-8 installed by default.

I'm taking the blunt approach here and ensuring in the pre-suite that on
almalinux glibc-all-langpacks is installed, and on debian locales-all is
installed.

The tweak to the reset_master_system_locale() helper, was my original
fix for the failures on debian, to ensure that we locale-gen en_US
before attempting to switch to it (since it's not there by default).
On Debian, the enable_local_language() method should have
been able to locale-gen desired locales after editing /etc/locale.gen,
but this was failing for the first calls to
  localectl set-locale LANG=en_US.utf8
in the enable_option_disable_i18n.rb test, althought not when done for
ja_JP, and not when run for en_US in other tests. I have no idea what
was going on there.