(gh-101) Fix up server_set_facts for serverimplementation
This test was failing because it was expecting a certain key ordering of
the $server_facts hash (that serverversion would be the first element),
and the addition of serverimplementation disrupted that.

Rather than trying to match on the printed order of the $server_facts,
changed to validate that we have a non-empty $server_facts hash.