(gh-101) Fix two tests to handle multiple agents
The Beaker agent() method is a dubious creature. It is generated, I
believe, by Beaker's Beaker::DSL::Roles.add_role_def() which TestCase
uses to generate helper methods from host role values to either an array
of matching host objects, or the singular host object if there is only
one. This later behavior makes these helpers dangerous since it
changes depending on whether Beaker is executed with say [agent1,
agent2] hosts or just agent1 hosts.

These two tests could pass if a single host was given, but would fail
when agent returned an array.

For the first case, confine works better than skip test since it
constrains the hosts array for the test case.

For the second case, the code is operating within an agents.each |host|
block, and it was probably a revision somewhere in its past that left
some methods stranded using agent() instead of host. It just happened to
work as expected in the single node case since agent() would return the
same value as host.