(gh-101) Skip i18n tests using deprecated function
The i18n tests are in an odd state. 3b82c0f6a85 added a skip_test to
acceptance/tests/i18n/modules/puppet_agent.rb because the
eputname-i18ndemo module used by the test has an init class with a call
to the old, deprecated stdlib validate_absolute_path() function that
would error out the test. However, puppet_agent_cached_ctalog,
puppet_apply, puppet_apply_module_lang and puppet_apply_unsupported_lang
are all in the same boat.

It's possible the test suite was configured to run with strict
disabled? Or the i18n/ tests were simply not being run, since there is
currently no language localization being done.

For now, I'm adding the skip_test to the other affected tests, and they
can be removed or the module can be updated in a future audit of the
suite.