Raise an error if the version is not changed
Either the version is already the one currently set (not something
supposed to happen, but not a big deal either, so we can fail), or the
files have been edited in a way the action fail to update the version
which is a major issue we do not want to see.

Detect if a version bump is moot and raise if it happens.