Make passing invalid options to install.rb fatal
In b909b673bae784af7822fcd58549fd6773b4320b the --[no-]prereq-checks
option was dropped. The error was printed, but then continued to execute
and it ignored all passed options. That caused a regression in 8.21.0.

It now catches the error and exits with a non-zero exit code, forcing
the packager to deal with it.