treat windows service accounts as case insensitive
Windows service accounts are *case preserving* but not *case sensitive*.

Before this change, Windows service logon account names were compared
directly as case sensitive. When there was a case mismatch, it was treated
as a change to the account and the service was restarted. This now
normalizes case before comparing, reducing spurious restarts.

Co-authored-by: Ewoud Kohl van Wijngaarden <ewoud@kohlvanwijngaarden.nl>