Move openvox-agent actions into main repo actions
This includes renaming the existing acceptance and making it clear which of these actions are for which things. This also removes the config:print check, because it's part of the 'packaging' gem, we'd have to move a bunch of old stuff from openvox-agent's Rakefile into openvox's, and we don't use it anyway.