Update components to pull version from version.rb
Rather than updating the version several different places, this utilizes the puppet/version library to get the version so we only have to updated it in version.rb. It's a little janky in the gemspec due to the gemspec being laid down on install in a different place, but this location should be stable.