Gate vanagon gem in Gemfile
Because Vanagon 1.0.0 requires Ruby >= 3.2, and because we do PR checks against JRuby 9.4 which is 3.1, and because the dependency resolver looks at *every* gem in the gemfile, whether the group is included or not, this gates the inclusion of the vanagon gem when on older Ruby versions. We don't really need the vanagon gem for anything but packaging.